added some error reports and made the size of the queue drop variable larger

+++ Pensel/firmware/include/LSM303DLHC.h
@@ -95,8 +95,8 @@ bool LSM303DLHC_mag_dataAvailable(void);
 ret_t LSM303DLHC_accel_getPacket(accel_packet_t * pkt_ptr, bool peak);
 ret_t LSM303DLHC_mag_getPacket(mag_packet_t * pkt_ptr, bool peak);
 // error counting methods...
-uint8_t LSM303DLHC_accel_packetOverwriteCount(void);
-uint8_t LSM303DLHC_mag_packetOverwriteCount(void);
+uint32_t LSM303DLHC_accel_packetOverwriteCount(void);
+uint32_t LSM303DLHC_mag_packetOverwriteCount(void);
 uint32_t LSM303DLHC_accel_HardwareOverwriteCount(void);
 uint32_t LSM303DLHC_mag_HardwareOverwriteCount(void);
